+++ src/filters/index.js
@@ -9,7 +9,7 @@
  * @namespace PIXI.filters
  */
 module.exports = {
-    //FXAAFilter:          require('./fxaa/FXAAFilter'),
+    FXAAFilter:          require('./fxaa_/FXAAFilter'),
     NoiseFilter:        require('./noise/NoiseFilter'),
     DisplacementFilter: require('./displacement/DisplacementFilter'),
     BlurFilter:         require('./blur/BlurFilter'),
